secciones terminadas

+++ server/components/s3/index.js
@@ -10,10 +10,15 @@ module.exports.uploadFile = function (files, type, cb) {
   var format
    ,  isArray = true;
 
+  // si no encuentra el tipo dentro del formato de imagenes debería retornar los archivos
   if (type === 'gallery'){
     format = images.gallery;
   } else if (type==='logo') {
     format = images.logo;
+  }else if(type==='largeSlider'){
+    format = images.largeSlider;
+  }else if(type==='mobileSlider'){
+    format = images.mobileSlider;
   }else{
     console.log('no se ha especificado tipo de imagen');
     return files;
@@ -28,11 +33,13 @@ module.exports.uploadFile = function (files, type, cb) {
 
   files.forEach(function (file, k) {
     Object.keys(format).forEach(function( el, i){
+      // no subas imagen  si es tipo personalizada
+      if (format[el].custom) return;
+
       var pathVariation = images.addSufixText(files[k].path, format[el].sufix);
       gm(files[k].path)
       .resize(format[el].width,format[el].width)
       .write(pathVariation, function (err) {
-        console.log('pathVariation->',pathVariation);
         if (!err) {
           uploadFile(pathVariation,pathVariation.split('/').slice(-1)[0],cb);
         }else{
@@ -43,6 +50,24 @@ module.exports.uploadFile = function (files, type, cb) {
   });
 
   return (isArray ? files : files[0]);
+}
+
+module.exports.oneUploadFile = function (file, obj,cb) {
+
+  if (!obj) {
+    return ;
+  }else{
+    if (!obj.kind) { return; }
+    if (!obj.size) { return; }
+  }
+  if (!file) { return; }
+
+  var kind = obj.kind;
+  var size = obj.size;
+  var sufix = images[kind][size].sufix;
+  var path = images.generateS3Url(file.name,sufix);
+  uploadFile(file.path,path.split('/').slice(-1)[0],cb);
+  return path;
 }
 
 module.exports.deleteFiles = function (images,name,cb) {
@@ -57,6 +82,10 @@ module.exports.deleteFiles = function (images,name,cb) {
   for (var i = images.length - 1; i >= 0; i--) {
     // por alguna razon json de mongoose no se puden hacer foreach
     // el siguiente código convierte json a string y los regresa a su forma original
+    console.log("to delete s3->", images[i]);
+    if (!images[i].paths) {
+      continue;
+    }
     images[i].paths = JSON.parse(JSON.stringify(images[i].paths));
     for(var k in images[i].paths){
       if( typeof images[i].paths[k] == 'string'  ){
@@ -67,18 +96,18 @@ module.exports.deleteFiles = function (images,name,cb) {
       }
     }
   };
-  console.log(objects);
-
-  var uploader = clientS3.deleteObjects({
-    Bucket: config.bucket,
-    Delete: {
-      Objects:objects
-    }
-  },function (err,data) {
-    if (err) console.log(err, err.stack); // an error occurred
-    else     console.log(data);           // successful response
-    if (cb) cb(err,data);
-  });
+  if (objects.length > 0) {
+    var uploader = clientS3.deleteObjects({
+      Bucket: config.bucket,
+      Delete: {
+        Objects:objects
+      }
+    },function (err,data) {
+      if (err) console.log(err, err.stack); // an error occurred
+      else     console.log(data);           // successful response
+      if (cb) cb(err,data);
+    });
+  }
 
   return (isArray ? images : images[0]);
 }
